+++ app/constants/colors.js
@@ -12,4 +12,4 @@ export const STATUS_COLORS = {
 };
 
 export const HEADER_BACKGROUND = isIOS ? '#FFF' : '#2F343D';
-export const HEADER_TINT = isIOS ? '#1d74f5' : '#FFF';
+export const HEADER_TITLE = isIOS ? '#0C0D0F' : '#FFF';
